# Numba matching speedup