--- a/asv.conf.json
+++ b/asv.conf.json
@@ -1,0 +1,11 @@
+{
+    "version": 1,
+    "project": "sparse",
+    "project_url": "https://sparse.pydata.org/",
+    "repo": ".",
+    "dvcs": "git",
+    "environment_type": "conda",
+    "env_dir": ".asv/env",
+    "results_dir": ".asv/results",
+    "html_dir": ".asv/html"
+}