--- a/requirements.txt
+++ b/requirements.txt
@@ -1,2 +1,3 @@
 numpy >= 1.13
 scipy >= 0.19
+numba