--- a/setup.py
+++ b/setup.py
@@ -26,7 +26,10 @@ extras_require = {
     ],
     'tests-cov': [
         'pytest-cov'
-    ]
+    ],
+    'bench': [
+        'asv'
+    ],
 }
 
 all_requires = []