--- a/sparse/coo/umath.py
+++ b/sparse/coo/umath.py
@@ -3,6 +3,8 @@ from itertools import product
 import numpy as np
 import scipy.sparse
 
+import numba
+
 from ..utils import isscalar, PositinalArgumentPartial, _zero_of_dtype
 from ..compatibility import range, zip, zip_longest
 
@@ -79,10 +81,8 @@ def elemwise(func, *args, **kwargs):
     return _elemwise_n_ary(func, *args, **kwargs)
 
 
-# (c) Paul Panzer
-# Taken from https://stackoverflow.com/a/47833496/774273
-# License: https://creativecommons.org/licenses/by-sa/3.0/
-def _match_arrays(a, b):
+@numba.jit(nopython=True)
+def _match_arrays(a, b):    # pragma: no cover
     """
     Finds all indexes into a and b such that a[i] = b[j]. The outputs are sorted
     in lexographical order.
@@ -99,29 +99,28 @@ def _match_arrays(a, b):
         The output indices of every possible pair of matching elements.
     """
     if len(a) == 0 or len(b) == 0:
-        return np.array([], dtype=np.uint8), np.array([], dtype=np.uint8)
-    asw = np.r_[0, 1 + np.flatnonzero(a[:-1] != a[1:]), len(a)]
-    bsw = np.r_[0, 1 + np.flatnonzero(b[:-1] != b[1:]), len(b)]
-    al, bl = np.diff(asw), np.diff(bsw)
-    na = len(al)
-    asw, bsw = asw, bsw
-    abunq = np.r_[a[asw[:-1]], b[bsw[:-1]]]
-    m = np.argsort(abunq, kind='mergesort')
-    mv = abunq[m]
-    midx = np.flatnonzero(mv[:-1] == mv[1:])
-    ai, bi = m[midx], m[midx + 1] - na
-    aic = np.r_[0, np.cumsum(al[ai])]
-    a_idx = np.ones((aic[-1],), dtype=np.int_)
-    a_idx[aic[:-1]] = asw[ai]
-    a_idx[aic[1:-1]] -= asw[ai[:-1]] + al[ai[:-1]] - 1
-    a_idx = np.repeat(np.cumsum(a_idx), np.repeat(bl[bi], al[ai]))
-    bi = np.repeat(bi, al[ai])
-    bic = np.r_[0, np.cumsum(bl[bi])]
-    b_idx = np.ones((bic[-1],), dtype=np.int_)
-    b_idx[bic[:-1]] = bsw[bi]
-    b_idx[bic[1:-1]] -= bsw[bi[:-1]] + bl[bi[:-1]] - 1
-    b_idx = np.cumsum(b_idx)
-    return a_idx, b_idx
+        return np.empty(0, dtype=np.uintp), np.empty(0, dtype=np.uintp)
+
+    a_ind, b_ind = [], []
+    nb = len(b)
+    ib = 0
+    match = 0
+
+    for ia, j in enumerate(a):
+        if j == b[match]:
+            ib = match
+
+        while ib < nb and j >= b[ib]:
+            if j == b[ib]:
+                a_ind.append(ia)
+                b_ind.append(ib)
+
+                if b[match] < b[ib]:
+                    match = ib
+
+            ib += 1
+
+    return np.array(a_ind, dtype=np.uintp), np.array(b_ind, dtype=np.uintp)
 
 
 def _elemwise_n_ary(func, *args, **kwargs):