Benchmarks

```
In[2]: import numpy as np
In[3]: import sparse
In[4]: x = sparse.random((1000, 1000, 1000), density=0.01)
In[5]: y = sparse.random((1000, 1000, 1000), density=0.01)
In[6]: x.sum_duplicates()
In[7]: y.sum_duplicates()
In[8]: a = x.linear_loc()
In[9]: b = y.linear_loc()
In[10]: %timeit sparse.coo.umath._match_arrays(a, b)
```

Old:
```
852 ms ± 13.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

New:
```
112 ms ± 2.21 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```