@rgommers 2018-03-17
How about making the benchmark code in this PR description into a proper benchmark? Copying the `asv` benchmark setup from scipy should be easy, and writing benchmarks is very simple.

About the `version.py` rename - my experience is that you want to keep that a private submodule.