@hameerabbasi 2018-03-17
Thanks for the input, @rgommers, it's very welcome. I've set up a `asv`. I'll hopefully keep adding benchmarks as time goes on.