@rgommers 2018-03-17
This looks good - not only an 8x speedup, but I find the code a lot more readable.