@hameerabbasi 2018-03-17
Waiting on @mrocklin's input before I merge - He usually has excellent insights into different aspects of everything.