@hameerabbasi 2018-03-18
A couple of replies. Yes, rebase-and-merge seems like a good idea.