@hameerabbasi 2018-03-18
cc @mrocklin If you could give this a final look before we merge, that'd be great. :-)