@mrocklin 2018-03-18
+1 from me.  Feel free to rebase and merge.  

One other question I have (which could be answered at another time if you'd like to go ahead with this now)  is where the extra missed lines are coming from.  Looking at the codecov CI I suspect that it may be coming from the benchmark files?  Is this correct?  It doesn't seem like it should given what you've written down in the codecov.yaml file, but I'm not as familiar with this, so I'm not sure.