@hameerabbasi 2018-03-18
The diff is here: https://codecov.io/gh/pydata/sparse/pull/127/diff#diff-YmVuY2htYXJrcy9iZW5jaG1hcmtfY29vLnB5

It seems to me that CodeCov uses the YAML from the base, it seems consistent with this and other behavior I've seen so far. This is unlike Travis, which uses the config in the PR.

There will be missed lines here but the next one should be fine, if my hunch is correct.

I validated the YAML accross their web API.